+++ R/print.R
@@ -10,8 +10,8 @@ print.incidence <- function(x, ...) {
                 sum(x$n), head(x$isoweeks, 1), tail(x$isoweeks, 1)))
   }
   if (ncol(x$counts) > 1L) {
-    groups.txt <- paste(colnames(x$counts), collapse = ", ")
-    cat(sprintf("[%d groups: %s]\n", ncol(x$counts), groups.txt))
+    groups.txt <- paste(group_names(x), collapse = ", ")
+    cat(sprintf("[%d groups: %s]\n", n_groups(x), groups.txt))
   }
   cat(sprintf("\n$counts: matrix with %d rows and %d columns\n",
               nrow(x$counts), ncol(x$counts)))